refactor: upgrade user ID to distributed Snowflake ID
- Add SnowflakeIdGenerator utility for generating global unique IDs
- Refactor LocalUser entity to use userId (Long) as primary key
- Update Order entity to reference userId (Long) instead of id (Integer)
- Update AuthService to use userId throughout
- Update LocalUserRepository to use Long as ID type
- Update AuthResponse DTO to use Long for id field
- Add migration script to refactor existing databases
- Add new schema-with-userid.sql for fresh deployments
- Add comprehensive USERID_REFACTORING_GUIDE.md documentation

Benefits:
- Supports distributed deployments and sharding
- Global unique IDs across multiple databases
- Trend-increasing IDs for better database performance
- Time information embedded in ID
- Better security (hard to enumerate)
- Supports 1024 nodes and 4096 IDs per millisecond

+++ server-java/src/main/java/com/usjobhunt/dto/AuthResponse.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class AuthResponse {
-    private Integer id;
+    private Long id;  // userId (Snowflake ID)
     private String email;
     private String name;
     private String token;
